refactor: main-view 프로젝트 정렬 Enum 및 로직 분리
